# update: removed all the pointers and indirection