--- a/rfc5424/parser.go
+++ b/rfc5424/parser.go
@@ -20,17 +20,10 @@ func NewParser() *Parser {
 // Parse parses the input RFC5424 syslog message using its FSM.
 //
 // Best effort mode enables the partial parsing.
-func (p *Parser) Parse(input []byte, bestEffort *bool) (*SyslogMessage, error) {
+func (p *Parser) Parse(input []byte, bestEffort bool) (msg SyslogMessage, err error) {
 	p.Lock()
 	defer p.Unlock()
 
-	msg, err := p.machine.Parse(input, bestEffort)
-	if err != nil {
-		if bestEffort != nil && *bestEffort != false {
-			return msg, err
-		}
-		return nil, err
-	}
-
-	return msg, nil
+	msg, err = p.machine.Parse(input, bestEffort)
+	return
 }