--- a/rfc5424/parser_test.go
+++ b/rfc5424/parser_test.go
@@ -13,12 +13,11 @@ func TestParserParse(t *testing.T) {
 		t.Run(rxpad(string(tc.input), 50), func(t *testing.T) {
 			t.Parallel()
 
-			bestEffort := true
-			message, merr := p.Parse(tc.input, nil)
-			partial, perr := p.Parse(tc.input, &bestEffort)
+			message, merr := p.Parse(tc.input, false)
+			partial, perr := p.Parse(tc.input, true)
 
 			if !tc.valid {
-				assert.Nil(t, message)
+				assert.Equal(t, message, SyslogMessage{})
 				assert.Error(t, merr)
 				assert.EqualError(t, merr, tc.errorString)