… rfc5424/machine.go
… rfc5424/machine_test.go