I admittedly don't understand the purpose of all the pointers and indirection in the code, but i did a little light profiling and then rolled it in our own system and found that the extra pointers were causing more stress on the allocator and GC.  I refactored the ragel system and some of the code to remove them and got between 10 and 60% speedups.  If there is some other reason for keeping that indirection I would like to better understand it.  I am also admittedly new to ragel and code generation, so if I botched some of the code changes let me know.

Thank you for the library, it was already fast, this is a little faster and more go-like.

```
benchmark                                                               old ns/op     new ns/op     delta
BenchmarkParse/[no]_empty_input__________________________________-4     270           233           -13.70%
BenchmarkParse/[no]_multiple_syslog_messages_on_multiple_lines___-4     519           322           -37.96%
BenchmarkParse/[no]_impossible_timestamp_________________________-4     1258          1065          -15.34%
BenchmarkParse/[no]_malformed_structured_data____________________-4     608           423           -30.43%
BenchmarkParse/[no]_with_duplicated_structured_data_id___________-4     1698          1397          -17.73%
BenchmarkParse/[ok]_minimal______________________________________-4     321           122           -61.99%
BenchmarkParse/[ok]_average_message______________________________-4     2588          2228          -13.91%
BenchmarkParse/[ok]_complicated_message__________________________-4     2176          1860          -14.52%
BenchmarkParse/[ok]_very_long_message____________________________-4     4501          4081          -9.33%
BenchmarkParse/[ok]_all_max_length_and_complete__________________-4     3514          3174          -9.68%
BenchmarkParse/[ok]_all_max_length_except_structured_data_and_mes-4     2115          1846          -12.72%
BenchmarkParse/[ok]_minimal_with_message_containing_newline______-4     384           157           -59.11%
BenchmarkParse/[ok]_w/o_procid,_w/o_structured_data,_with_message-4     1082          762           -29.57%
BenchmarkParse/[ok]_minimal_with_UTF-8_message___________________-4     608           370           -39.14%
BenchmarkParse/[ok]_with_structured_data_id,_w/o_structured_data_-4     1231          950           -22.83%
BenchmarkParse/[ok]_with_multiple_structured_data________________-4     1920          1603          -16.51%
BenchmarkParse/[ok]_with_escaped_backslash_within_structured_data-4     1667          1319          -20.88%
BenchmarkParse/[ok]_with_UTF-8_structured_data_param_value,_with_-4     1813          1596          -11.97%
```