@leodido 2018-04-24
Hi @traetox first of all thanks for your contribution.

I gavea quick look at it. Anyway after I'll look better at it.

But two considerations:
- the part about `output` variable is probably correct
- removing the pointers from the `SyslogMessage` fields is probably not want we want since we cannot represent the nil values of the RFC with the zero-values of golang types (except that for the version part)

Can you please reword your commit following the repository's convention?
Something like `Update: Removing pointers from ...`