@traetox 2018-04-24
I am going to rework the changes to incorporate an unexported flag that indicates whether or not the SyslogMessage is set.  Accessor methods will then return the value and an ok which probably breaks your usage of the library.  As such, I am going to close the PR.