@leodido 2018-04-27
Hey @traetox have a look at PR #7.

Now avoiding pointers & indirections as you suggested, but without chaning the interface at all (using an intermediate object and some flags).